Rename distributional assertions module to distribution
